requirements.txt
Levenshtein==0.20.9
Pillow==9.3.0
rapidfuzz==2.13.7
